requirements.txt
streamlit
pandas
requests
python-dotenv
matplotlib
koreanize-matplotlib
